fix(component): Clean up results compoent   * Needs a bit more work for display resutls   * Should display resutls after tests are run   * If not running then show last run restuls
