+++ repos/tap/components/tapIndex.js
@@ -5,6 +5,7 @@ export * from 'SVComponents/chip'
 export * from 'SVComponents/domStyles'
 export * from 'SVComponents/feature'
 export * from 'SVComponents/list'
+export * from 'SVComponents//results'
 export * from 'SVComponents/step'
 export * from 'SVComponents/scenario'
 export * from 'SVComponents/sidebar'
